chore(assembly): split out AssemblerError/Product into new modules
